refactor: move database to shared folder between services
